LayoutTests/fast/selectors/lang-valid-extended-filtering-expected.txt

Verify selector specifying extended filtering of :lang() pseudo class

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


- af-ZA
PASS document.querySelectorAll(':lang(af)').length == 1 is true
PASS document.querySelectorAll(':lang(af-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(af-ZA)').length == 1 is true
PASS document.querySelectorAll(':lang(af-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ZA")').length == 1 is true

- ar-AE
PASS document.querySelectorAll(':lang(ar)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-AE)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-AE")').length == 1 is true

- ar-BH
PASS document.querySelectorAll(':lang(ar)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-BH)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BH")').length == 1 is true

- ar-YE
PASS document.querySelectorAll(':lang(ar)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-YE)').length == 1 is true
PASS document.querySelectorAll(':lang(ar-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-YE")').length == 1 is true

- art-lojban
PASS document.querySelectorAll(':lang(art)').length == 1 is true
PASS document.querySelectorAll(':lang(art-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(art-lojban)').length == 1 is true
PASS document.querySelectorAll(':lang(art-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-lojban")').length == 1 is true

- az-Arab-IR
PASS document.querySelectorAll(':lang(az)').length == 1 is true
PASS document.querySelectorAll(':lang(az-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(az-Arab)').length == 1 is true
PASS document.querySelectorAll(':lang(az-\\*-IR)').length == 1 is true
PASS document.querySelectorAll(':lang(az-Arab-IR)').length == 1 is true
PASS document.querySelectorAll(':lang(az-Arab-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Arab-IR")').length == 1 is true

- be-BY
PASS document.querySelectorAll(':lang(be)').length == 1 is true
PASS document.querySelectorAll(':lang(be-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(be-BY)').length == 1 is true
PASS document.querySelectorAll(':lang(be-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BY")').length == 1 is true

- bg-BG
PASS document.querySelectorAll(':lang(bg)').length == 1 is true
PASS document.querySelectorAll(':lang(bg-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(bg-BG)').length == 1 is true
PASS document.querySelectorAll(':lang(bg-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BG")').length == 1 is true

- ca-ES
PASS document.querySelectorAll(':lang(ca)').length == 1 is true
PASS document.querySelectorAll(':lang(ca-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(ca-ES)').length == 1 is true
PASS document.querySelectorAll(':lang(ca-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ES")').length == 1 is true

- cs-CZ
PASS document.querySelectorAll(':lang(cs)').length == 1 is true
PASS document.querySelectorAll(':lang(cs-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(cs-CZ)').length == 1 is true
PASS document.querySelectorAll(':lang(cs-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CZ")').length == 1 is true

- cy-GB
PASS document.querySelectorAll(':lang(cy)').length == 1 is true
PASS document.querySelectorAll(':lang(cy-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(cy-GB)').length == 1 is true
PASS document.querySelectorAll(':lang(cy-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-GB")').length == 1 is true

- de-AT
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-AT)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-AT")').length == 1 is true

- de-CH
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH")').length == 1 is true

- de-CH-1996
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-1996)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH-1996)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH-1996")').length == 1 is true

- de-CH-1997
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-1997)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH-1997)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH-1997")').length == 1 is true

- de-DE
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-DE")').length == 1 is true

- de-DE-1996
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-1996)').length == 1 is true
PASS document.querySelectorAll(':lang(de-DE-1996)').length == 1 is true
PASS document.querySelectorAll(':lang(de-DE-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-DE-1996")').length == 1 is true

- de-Latn-DE
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Latn-DE")').length == 1 is true

- de-Latf-DE
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latf)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latf-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latf-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Latf-DE")').length == 1 is true

- de-Latn-DE-1996
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*-DE-1996)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn-\\*-1996)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn-DE-1996)').length == 1 is true
PASS document.querySelectorAll(':lang(de-Latn-DE-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Latn-DE-1996")').length == 1 is true

- en-US
PASS document.querySelectorAll(':lang(en)').length == 1 is true
PASS document.querySelectorAll(':lang(en-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(en-US)').length == 1 is true
PASS document.querySelectorAll(':lang(en-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-US")').length == 1 is true

- en-GB
PASS document.querySelectorAll(':lang(en)').length == 1 is true
PASS document.querySelectorAll(':lang(en-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(en-GB)').length == 1 is true
PASS document.querySelectorAll(':lang(en-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-GB")').length == 1 is true

- en-Latn-Brai
PASS document.querySelectorAll(':lang(en)').length == 1 is true
PASS document.querySelectorAll(':lang(en-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(en-Latn)').length == 1 is true
PASS document.querySelectorAll(':lang(en-\\*-Brai)').length == 1 is true
PASS document.querySelectorAll(':lang(en-Latn-Brai)').length == 1 is true
PASS document.querySelectorAll(':lang(en-Latn-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Latn-Brai")').length == 1 is true

- es-ES
PASS document.querySelectorAll(':lang(es)').length == 1 is true
PASS document.querySelectorAll(':lang(es-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(es-ES)').length == 1 is true
PASS document.querySelectorAll(':lang(es-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ES")').length == 1 is true

- es-ES-valencia
PASS document.querySelectorAll(':lang(es)').length == 1 is true
PASS document.querySelectorAll(':lang(es-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(es-ES)').length == 1 is true
PASS document.querySelectorAll(':lang(es-\\*-valencia)').length == 1 is true
PASS document.querySelectorAll(':lang(es-ES-valencia)').length == 1 is true
PASS document.querySelectorAll(':lang(es-ES-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ES-valencia")').length == 1 is true

- es-AR
PASS document.querySelectorAll(':lang(es)').length == 1 is true
PASS document.querySelectorAll(':lang(es-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(es-AR)').length == 1 is true
PASS document.querySelectorAll(':lang(es-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-AR")').length == 1 is true

- fr-BE
PASS document.querySelectorAll(':lang(fr)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-BE)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BE")').length == 1 is true

- fr-CA
PASS document.querySelectorAll(':lang(fr)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-CA)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CA")').length == 1 is true

- fr-CH
PASS document.querySelectorAll(':lang(fr)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH")').length == 1 is true

- fr-FR
PASS document.querySelectorAll(':lang(fr)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-FR)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-FR")').length == 1 is true

- fr-LU
PASS document.querySelectorAll(':lang(fr)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-LU)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-LU")').length == 1 is true

- fr-MC
PASS document.querySelectorAll(':lang(fr)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-MC)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-MC")').length == 1 is true

- gl-ES
PASS document.querySelectorAll(':lang(gl)').length == 1 is true
PASS document.querySelectorAll(':lang(gl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(gl-ES)').length == 1 is true
PASS document.querySelectorAll(':lang(gl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ES")').length == 1 is true

- gu-IN
PASS document.querySelectorAll(':lang(gu)').length == 1 is true
PASS document.querySelectorAll(':lang(gu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(gu-IN)').length == 1 is true
PASS document.querySelectorAll(':lang(gu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-IN")').length == 1 is true

- hr-BA
PASS document.querySelectorAll(':lang(hr)').length == 1 is true
PASS document.querySelectorAll(':lang(hr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(hr-BA)').length == 1 is true
PASS document.querySelectorAll(':lang(hr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BA")').length == 1 is true

- id-ID
PASS document.querySelectorAll(':lang(id)').length == 1 is true
PASS document.querySelectorAll(':lang(id-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(id-ID)').length == 1 is true
PASS document.querySelectorAll(':lang(id-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ID")').length == 1 is true

- i-klingon
PASS document.querySelectorAll(':lang(i)').length == 1 is true
PASS document.querySelectorAll(':lang(i-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(i-klingon)').length == 1 is true
PASS document.querySelectorAll(':lang(i-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-klingon")').length == 1 is true

- ja-JP
PASS document.querySelectorAll(':lang(ja)').length == 1 is true
PASS document.querySelectorAll(':lang(ja-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(ja-JP)').length == 1 is true
PASS document.querySelectorAll(':lang(ja-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-JP")').length == 1 is true

- ko-KR
PASS document.querySelectorAll(':lang(ko)').length == 1 is true
PASS document.querySelectorAll(':lang(ko-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(ko-KR)').length == 1 is true
PASS document.querySelectorAll(':lang(ko-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-KR")').length == 1 is true

- lt-LT
PASS document.querySelectorAll(':lang(lt)').length == 1 is true
PASS document.querySelectorAll(':lang(lt-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(lt-LT)').length == 1 is true
PASS document.querySelectorAll(':lang(lt-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-LT")').length == 1 is true

- mi-NZ
PASS document.querySelectorAll(':lang(mi)').length == 1 is true
PASS document.querySelectorAll(':lang(mi-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(mi-NZ)').length == 1 is true
PASS document.querySelectorAll(':lang(mi-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-NZ")').length == 1 is true

- mn-MN
PASS document.querySelectorAll(':lang(mn)').length == 1 is true
PASS document.querySelectorAll(':lang(mn-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(mn-MN)').length == 1 is true
PASS document.querySelectorAll(':lang(mn-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-MN")').length == 1 is true

- nl-BE
PASS document.querySelectorAll(':lang(nl)').length == 1 is true
PASS document.querySelectorAll(':lang(nl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(nl-BE)').length == 1 is true
PASS document.querySelectorAll(':lang(nl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BE")').length == 1 is true

- nl-NL
PASS document.querySelectorAll(':lang(nl)').length == 1 is true
PASS document.querySelectorAll(':lang(nl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(nl-NL)').length == 1 is true
PASS document.querySelectorAll(':lang(nl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-NL")').length == 1 is true

- pl-PL
PASS document.querySelectorAll(':lang(pl)').length == 1 is true
PASS document.querySelectorAll(':lang(pl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(pl-PL)').length == 1 is true
PASS document.querySelectorAll(':lang(pl-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-PL")').length == 1 is true

- qu-BO
PASS document.querySelectorAll(':lang(qu)').length == 1 is true
PASS document.querySelectorAll(':lang(qu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(qu-BO)').length == 1 is true
PASS document.querySelectorAll(':lang(qu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BO")').length == 1 is true

- se-FI
PASS document.querySelectorAll(':lang(se)').length == 1 is true
PASS document.querySelectorAll(':lang(se-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(se-FI)').length == 1 is true
PASS document.querySelectorAll(':lang(se-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-FI")').length == 1 is true

- se-NO
PASS document.querySelectorAll(':lang(se)').length == 1 is true
PASS document.querySelectorAll(':lang(se-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(se-NO)').length == 1 is true
PASS document.querySelectorAll(':lang(se-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-NO")').length == 1 is true

- sgn-BE-FR
PASS document.querySelectorAll(':lang(sgn)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-BE)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-\\*-FR)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-BE-FR)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-BE-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BE-FR")').length == 1 is true

- sgn-BE-NL
PASS document.querySelectorAll(':lang(sgn)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-BE)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-\\*-NL)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-BE-NL)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-BE-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-BE-NL")').length == 1 is true

- sgn-CH-DE
PASS document.querySelectorAll(':lang(sgn)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-\\*-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-CH-DE)').length == 1 is true
PASS document.querySelectorAll(':lang(sgn-CH-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH-DE")').length == 1 is true

- sr-Latn-RS
PASS document.querySelectorAll(':lang(sr)').length == 1 is true
PASS document.querySelectorAll(':lang(sr-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(sr-Latn)').length == 1 is true
PASS document.querySelectorAll(':lang(sr-\\*-RS)').length == 1 is true
PASS document.querySelectorAll(':lang(sr-Latn-RS)').length == 1 is true
PASS document.querySelectorAll(':lang(sr-Latn-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Latn-RS")').length == 1 is true

- sv-SE
PASS document.querySelectorAll(':lang(sv)').length == 1 is true
PASS document.querySelectorAll(':lang(sv-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(sv-SE)').length == 1 is true
PASS document.querySelectorAll(':lang(sv-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-SE")').length == 1 is true

- syr-SY
PASS document.querySelectorAll(':lang(syr)').length == 1 is true
PASS document.querySelectorAll(':lang(syr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(syr-SY)').length == 1 is true
PASS document.querySelectorAll(':lang(syr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-SY")').length == 1 is true

- tt-RU
PASS document.querySelectorAll(':lang(tt)').length == 1 is true
PASS document.querySelectorAll(':lang(tt-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(tt-RU)').length == 1 is true
PASS document.querySelectorAll(':lang(tt-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-RU")').length == 1 is true

- uz-UZ
PASS document.querySelectorAll(':lang(uz)').length == 1 is true
PASS document.querySelectorAll(':lang(uz-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(uz-UZ)').length == 1 is true
PASS document.querySelectorAll(':lang(uz-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-UZ")').length == 1 is true

- vi-VN
PASS document.querySelectorAll(':lang(vi)').length == 1 is true
PASS document.querySelectorAll(':lang(vi-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(vi-VN)').length == 1 is true
PASS document.querySelectorAll(':lang(vi-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-VN")').length == 1 is true

- xh-ZA
PASS document.querySelectorAll(':lang(xh)').length == 1 is true
PASS document.querySelectorAll(':lang(xh-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(xh-ZA)').length == 1 is true
PASS document.querySelectorAll(':lang(xh-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ZA")').length == 1 is true

- yue-Hant-HK
PASS document.querySelectorAll(':lang(yue)').length == 1 is true
PASS document.querySelectorAll(':lang(yue-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(yue-Hant)').length == 1 is true
PASS document.querySelectorAll(':lang(yue-\\*-HK)').length == 1 is true
PASS document.querySelectorAll(':lang(yue-Hant-HK)').length == 1 is true
PASS document.querySelectorAll(':lang(yue-Hant-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Hant-HK")').length == 1 is true

- zh-yue-Hant-HK
PASS document.querySelectorAll(':lang(zh)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-yue)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*-Hant-HK)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-yue-Hant)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-yue-\\*-HK)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-yue-Hant-HK)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-yue-Hant-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-yue-Hant-HK")').length == 1 is true

- zh-CN
PASS document.querySelectorAll(':lang(zh)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-CN)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CN")').length == 1 is true

- zh-Hant-CN
PASS document.querySelectorAll(':lang(zh)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-Hant)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*-CN)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-Hant-CN)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-Hant-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Hant-CN")').length == 1 is true

- zh-Hans-CN
PASS document.querySelectorAll(':lang(zh)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-Hans)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*-CN)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-Hans-CN)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-Hans-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Hans-CN")').length == 1 is true

- zh-Hant
PASS document.querySelectorAll(':lang(zh)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-Hant)').length == 1 is true
PASS document.querySelectorAll(':lang(zh-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-Hant")').length == 1 is true

- de-CH
PASS document.querySelectorAll(':lang(de)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(de-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(de-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH")').length == 1 is true

- it-CH
PASS document.querySelectorAll(':lang(it)').length == 1 is true
PASS document.querySelectorAll(':lang(it-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(it-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(it-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH")').length == 2 is true

- fr-CH
PASS document.querySelectorAll(':lang(fr)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(fr-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH")').length == 3 is true

- rm-CH
PASS document.querySelectorAll(':lang(rm)').length == 1 is true
PASS document.querySelectorAll(':lang(rm-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(rm-CH)').length == 1 is true
PASS document.querySelectorAll(':lang(rm-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-CH")').length == 4 is true

- de-CH,it-CH,fr-CH,rm-CH
PASS document.querySelectorAll(':lang("*-CH")').length == 4 is true

- hi-IN
PASS document.querySelectorAll(':lang(hi)').length == 1 is true
PASS document.querySelectorAll(':lang(hi-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(hi-IN)').length == 1 is true
PASS document.querySelectorAll(':lang(hi-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-IN")').length == 1 is true

- gu-IN
PASS document.querySelectorAll(':lang(gu)').length == 1 is true
PASS document.querySelectorAll(':lang(gu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(gu-IN)').length == 1 is true
PASS document.querySelectorAll(':lang(gu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-IN")').length == 2 is true

- kok-IN
PASS document.querySelectorAll(':lang(kok)').length == 1 is true
PASS document.querySelectorAll(':lang(kok-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(kok-IN)').length == 1 is true
PASS document.querySelectorAll(':lang(kok-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-IN")').length == 3 is true

- hi-IN,gu-IN,kok-IN
PASS document.querySelectorAll(':lang("*-IN")').length == 3 is true

- xa-ZA
PASS document.querySelectorAll(':lang(xa)').length == 1 is true
PASS document.querySelectorAll(':lang(xa-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(xa-ZA)').length == 1 is true
PASS document.querySelectorAll(':lang(xa-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ZA")').length == 1 is true

- zu-ZA
PASS document.querySelectorAll(':lang(zu)').length == 1 is true
PASS document.querySelectorAll(':lang(zu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(zu-ZA)').length == 1 is true
PASS document.querySelectorAll(':lang(zu-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-ZA")').length == 2 is true

- xa-ZA,zu-ZA
PASS document.querySelectorAll(':lang("*-ZA")').length == 2 is true

- se-FI
PASS document.querySelectorAll(':lang(se)').length == 1 is true
PASS document.querySelectorAll(':lang(se-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(se-FI)').length == 1 is true
PASS document.querySelectorAll(':lang(se-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-FI")').length == 1 is true

- sv-FI
PASS document.querySelectorAll(':lang(sv)').length == 1 is true
PASS document.querySelectorAll(':lang(sv-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(sv-FI)').length == 1 is true
PASS document.querySelectorAll(':lang(sv-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-FI")').length == 2 is true

- se-FI,sv-FI
PASS document.querySelectorAll(':lang("*-FI")').length == 2 is true

- ur-PK
PASS document.querySelectorAll(':lang(ur)').length == 1 is true
PASS document.querySelectorAll(':lang(ur-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang(ur-PK)').length == 1 is true
PASS document.querySelectorAll(':lang(ur-\\*)').length == 1 is true
PASS document.querySelectorAll(':lang("*-PK")').length == 1 is true

- ur-PK
PASS document.querySelectorAll(':lang("*-PK")').length == 1 is true

PASS successfullyParsed is true

TEST COMPLETE
